refactor: align library modules with domain boundaries
