+++ src/components/transport.tsx
@@ -11,9 +11,9 @@ import { useRef, useState } from "react";
 import { useDraftInput } from "../hooks/use-draft-input";
 import { useAudioStatus, useTransport } from "../hooks/use-transport";
 import { useWindowEvent } from "../hooks/use-window-event";
-import { audioManager, GM_PROGRAMS } from "../lib/audio";
+import { audioManager } from "../lib/audio";
+import { GM_PROGRAMS } from "../lib/general-midi";
 import { isShortcutTextInputTarget, matchKeyboardEvent } from "../lib/keyboard";
-import { cn } from "../lib/utils";
 import { useProjectStore } from "../stores/project-store";
 import { COMMON_TIME_SIGNATURES, type GridSnap } from "../types";
 import { Button } from "./ui/button";
@@ -33,6 +33,7 @@ import {
   DropdownMenuTrigger,
 } from "./ui/dropdown-menu";
 import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
+import { cn } from "./ui/utils";
 
 function formatTimeCompact(seconds: number): string {
   const mins = Math.floor(seconds / 60);
