+++ src/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../lib/utils";
+import { cn } from "./utils";
 
 const buttonClassName =
   "inline-flex items-center justify-center rounded-md border border-border transition-colors focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] outline-none disabled:pointer-events-none disabled:opacity-50";
