+++ src/components/ui/popover.tsx
@@ -1,6 +1,6 @@
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 import * as React from "react";
-import { cn } from "../../lib/utils";
+import { cn } from "./utils";
 
 export function Popover({
   ...props
