+++ src/components/ui/toggle.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../lib/utils";
+import { cn } from "./utils";
 
 type ToggleProps = Omit<
   React.ComponentProps<"button">,
