+++ src/lib/audio.ts
@@ -1,162 +1,12 @@
-import { toast } from "sonner";
 import * as Tone from "tone";
 import oxisynthWasmUrl from "../assets/oxisynth/oxisynth.wasm?url";
 import oxisynthWorkletUrl from "../assets/oxisynth/worklet.js?url";
 import soundfontUrl from "../assets/soundfonts/A320U.sf2?url";
 import type { AudioTrack, ProjectState } from "../stores/project-store";
 import type { Note } from "../types";
-import { type AudioView, createAudioView } from "./audio-view";
 import { Metronome } from "./metronome";
+import { clampGain } from "./music";
 import { OxiSynthSynth } from "./oxisynth-synth";
-import { clampGain } from "./volume";
-
-// General MIDI Program Names (0-127)
-export const GM_PROGRAMS = [
-  // Piano (0-7)
-  "Acoustic Grand Piano",
-  "Bright Acoustic Piano",
-  "Electric Grand Piano",
-  "Honky-tonk Piano",
-  "Electric Piano 1",
-  "Electric Piano 2",
-  "Harpsichord",
-  "Clavinet",
-  // Chromatic Percussion (8-15)
-  "Celesta",
-  "Glockenspiel",
-  "Music Box",
-  "Vibraphone",
-  "Marimba",
-  "Xylophone",
-  "Tubular Bells",
-  "Dulcimer",
-  // Organ (16-23)
-  "Drawbar Organ",
-  "Percussive Organ",
-  "Rock Organ",
-  "Church Organ",
-  "Reed Organ",
-  "Accordion",
-  "Harmonica",
-  "Tango Accordion",
-  // Guitar (24-31)
-  "Acoustic Guitar (nylon)",
-  "Acoustic Guitar (steel)",
-  "Electric Guitar (jazz)",
-  "Electric Guitar (clean)",
-  "Electric Guitar (muted)",
-  "Overdriven Guitar",
-  "Distortion Guitar",
-  "Guitar Harmonics",
-  // Bass (32-39)
-  "Acoustic Bass",
-  "Electric Bass (finger)",
-  "Electric Bass (pick)",
-  "Fretless Bass",
-  "Slap Bass 1",
-  "Slap Bass 2",
-  "Synth Bass 1",
-  "Synth Bass 2",
-  // Strings (40-47)
-  "Violin",
-  "Viola",
-  "Cello",
-  "Contrabass",
-  "Tremolo Strings",
-  "Pizzicato Strings",
-  "Orchestral Harp",
-  "Timpani",
-  // Ensemble (48-55)
-  "String Ensemble 1",
-  "String Ensemble 2",
-  "Synth Strings 1",
-  "Synth Strings 2",
-  "Choir Aahs",
-  "Voice Oohs",
-  "Synth Voice",
-  "Orchestra Hit",
-  // Brass (56-63)
-  "Trumpet",
-  "Trombone",
-  "Tuba",
-  "Muted Trumpet",
-  "French Horn",
-  "Brass Section",
-  "Synth Brass 1",
-  "Synth Brass 2",
-  // Reed (64-71)
-  "Soprano Sax",
-  "Alto Sax",
-  "Tenor Sax",
-  "Baritone Sax",
-  "Oboe",
-  "English Horn",
-  "Bassoon",
-  "Clarinet",
-  // Pipe (72-79)
-  "Piccolo",
-  "Flute",
-  "Recorder",
-  "Pan Flute",
-  "Blown Bottle",
-  "Shakuhachi",
-  "Whistle",
-  "Ocarina",
-  // Synth Lead (80-87)
-  "Lead 1 (square)",
-  "Lead 2 (sawtooth)",
-  "Lead 3 (calliope)",
-  "Lead 4 (chiff)",
-  "Lead 5 (charang)",
-  "Lead 6 (voice)",
-  "Lead 7 (fifths)",
-  "Lead 8 (bass + lead)",
-  // Synth Pad (88-95)
-  "Pad 1 (new age)",
-  "Pad 2 (warm)",
-  "Pad 3 (polysynth)",
-  "Pad 4 (choir)",
-  "Pad 5 (bowed)",
-  "Pad 6 (metallic)",
-  "Pad 7 (halo)",
-  "Pad 8 (sweep)",
-  // Synth Effects (96-103)
-  "FX 1 (rain)",
-  "FX 2 (soundtrack)",
-  "FX 3 (crystal)",
-  "FX 4 (atmosphere)",
-  "FX 5 (brightness)",
-  "FX 6 (goblins)",
-  "FX 7 (echoes)",
-  "FX 8 (sci-fi)",
-  // Ethnic (104-111)
-  "Sitar",
-  "Banjo",
-  "Shamisen",
-  "Koto",
-  "Kalimba",
-  "Bagpipe",
-  "Fiddle",
-  "Shanai",
-  // Percussive (112-119)
-  "Tinkle Bell",
-  "Agogo",
-  "Steel Drums",
-  "Woodblock",
-  "Taiko Drum",
-  "Melodic Tom",
-  "Synth Drum",
-  "Reverse Cymbal",
-  // Sound Effects (120-127)
-  "Guitar Fret Noise",
-  "Breath Noise",
-  "Seashore",
-  "Bird Tweet",
-  "Telephone Ring",
-  "Helicopter",
-  "Applause",
-  "Gunshot",
-] as const;
 
 // Synth/scheduling readiness. The editor mounts before (and regardless of)
 // audio being usable, so playback capability is explicit state, not an
@@ -515,60 +365,4 @@ export function unlockAudioOnFirstGesture(): void {
   };
   window.addEventListener("pointerdown", unlock, true);
   window.addEventListener("keydown", unlock, true);
-}
-
-// Derive POINTS_PER_SECOND from max zoom level we want to support
-// Goal: 1 point per pixel at max zoom
-//
-// At 100 BPM, 4 beats visible (1 bar) in 1920px viewport:
-//   duration = 4 beats × 0.6 sec/beat = 2.4 sec
-//   points needed = 1920 px / 2.4 sec = 800 points/sec
-//
-// Formula: POINTS_PER_SECOND = viewportWidth / (beatsAtMaxZoom × secPerBeat)
-// With viewportWidth=1920, beatsAtMaxZoom=4, BPM=100:
-//   = 1920 / (4 × 60/100) = 1920 / 2.4 = 800
-const POINTS_PER_SECOND = 800;
-
-// Bailout threshold: avoid freezing on very long audio
-// At 48kHz, 10 min = 28.8M samples → ~50-100ms extraction (acceptable)
-// At 48kHz, 60 min = 172.8M samples → ~300-600ms extraction (too slow)
-const MAX_AUDIO_DURATION_SECONDS = 600; // 10 minutes
-
-// Load audio file and create AudioView for waveform display.
-// audioView is undefined when the waveform is skipped (too-long bailout); callers
-// map that to AudioWaveform "unavailable".
-export async function loadAudioFile(file: File): Promise<{
-  buffer: Tone.ToneAudioBuffer;
-  audioView?: AudioView;
-  duration: number;
-}> {
-  const url = URL.createObjectURL(file);
-  try {
-    const buffer = await Tone.ToneAudioBuffer.fromUrl(url);
-
-    // Bailout for very long audio
-    if (buffer.duration > MAX_AUDIO_DURATION_SECONDS) {
-      toast.warning(
-        `Audio too long (${Math.round(buffer.duration / 60)} min). Waveform disabled.`,
-      );
-      return {
-        buffer,
-        duration: buffer.duration,
-      };
-    }
-
-    const samples = buffer.getChannelData(0); // Use left/mono channel
-    const audioView = createAudioView(
-      samples,
-      buffer.sampleRate,
-      POINTS_PER_SECOND,
-    );
-    return {
-      buffer,
-      audioView,
-      duration: buffer.duration,
-    };
-  } finally {
-    URL.revokeObjectURL(url);
-  }
 }
